requirements.txt
numpy>=1.26
PyYAML>=6.0
soundfile>=0.12
torch>=2.8
torchaudio>=2.8
transformers>=4.57

